+++ src/Ad/InLine.php
@@ -2,15 +2,11 @@
 
 namespace Sokil\Vast\Ad;
 
+use Sokil\Vast\Creative\AbstractLinearCreative;
 use Sokil\Vast\Creative\InLine\Linear;
 
 class InLine extends AbstractAdNode
 {
-    /**
-     * @var \DomElement
-     */
-    private $extensionsDomElement;
-    
     /**
      * Set Ad title
      *
@@ -21,6 +17,7 @@ class InLine extends AbstractAdNode
     public function setAdTitle($value)
     {
         $this->setScalarNodeCdata('AdTitle', $value);
+
         return $this;
     }
 
@@ -35,44 +32,11 @@ class InLine extends AbstractAdNode
     /**
      * Create Linear creative
      *
-     * @return Linear
+     * @throws \Exception
+     * @return AbstractLinearCreative
      */
     public function createLinearCreative()
     {
         return $this->buildCreative('Linear');
-    }
-
-    /**
-     * Add extension
-     *
-     * @param string $type
-     * @param string $value
-     *
-     * @return $this
-     */
-    public function addExtension($type, $value)
-    {
-        // get container
-        if (!$this->extensionsDomElement) {
-            // get creatives tag
-            $this->extensionsDomElement = $this->getDomElement()->getElementsByTagName('Extensions')->item(0);
-            if (!$this->extensionsDomElement) {
-                $this->extensionsDomElement = $this->getDomElement()->ownerDocument->createElement('Extensions');
-                $this->getDomElement()->firstChild->appendChild($this->extensionsDomElement);
-            }
-        }
-        
-        // Creative dom element
-        $extensionDomElement = $this->extensionsDomElement->ownerDocument->createElement('Extension');
-        $this->extensionsDomElement->appendChild($extensionDomElement);
-        
-        // create cdata
-        $cdata = $this->getDomElement()->ownerDocument->createCDATASection($value);
-        
-        // append
-        $extensionDomElement->setAttribute('type', $type);
-        $extensionDomElement->appendChild($cdata);
-        
-        return $this;
     }
 }