+++ tests/FactoryTest.php
@@ -4,31 +4,41 @@ namespace Sokil\Vast;
 
 class FactoryTest extends AbstractTestCase
 {
-    /**
-     * @var string
-     */
-    private $vastXml = '<?xml version="1.0" encoding="UTF-8"?><VAST version="2.0"><Ad id="ad1"><InLine><AdSystem><![CDATA[Ad Server Name]]></AdSystem><AdTitle><![CDATA[Ad Title]]></AdTitle><Impression><![CDATA[http://ad.server.com/impression]]></Impression><Creatives><Creative><Linear><Duration>00:02:08</Duration><VideoClicks><ClickThrough><![CDATA[http://entertainmentserver.com/landing]]></ClickThrough><ClickTracking><![CDATA[http://ad.server.com/videoclicks/clicktracking]]></ClickTracking><CustomClick><![CDATA[http://ad.server.com/videoclicks/customclick]]></CustomClick></VideoClicks><TrackingEvents><Tracking event="start"><![CDATA[http://ad.server.com/trackingevent/start]]></Tracking><Tracking event="pause"><![CDATA[http://ad.server.com/trackingevent/stop]]></Tracking></TrackingEvents><MediaFiles><MediaFile delivery="progressive" type="video/mp4" height="100" width="100"><![CDATA[http://server.com/media.mp4]]></MediaFile></MediaFiles></Linear></Creative></Creatives></InLine></Ad></VAST>';
-
     public function testFromFile()
     {
-        $tempFilename = tempnam(sys_get_temp_dir(), "vast");
-        file_put_contents($tempFilename, $this->vastXml);
-
         $factory = new Factory();
-        $vastDocument = $factory->fromFile($tempFilename);
+        $vastDocument = $factory->fromFile(__DIR__ . '/vast.xml');
 
+        // check if loaded
         $this->assertInstanceOf(
             'Sokil\Vast\Document',
             $vastDocument
         );
 
-        unlink($tempFilename);
+        // get first ad section
+        $adSections = $vastDocument->getAdSections();
+        $adSection = $adSections[0];
+
+        // get scalar node
+        $adSystem = $adSection->getAdSystem();
+        $this->assertSame('Ad Server Name', $adSystem);
+
+        // get multi-nodes
+
+        $this->assertEquals(
+            array(
+                'http://ad.server.com/impression1',
+                'http://ad.server.com/impression2',
+                'http://ad.server.com/impression3',
+            ),
+            $adSection->getImpressions()
+        );
     }
 
     public function testFromString()
     {
         $factory = new Factory();
-        $vastDocument = $factory->fromString($this->vastXml);
+        $vastDocument = $factory->fromString(file_get_contents(__DIR__ . '/vast.xml'));
 
         $this->assertInstanceOf(
             'Sokil\Vast\Document',